+++ src/functions/generate-numbers-array.ts
@@ -1,3 +1,17 @@
-export function generateNumbersArray(length: number): number[] {
-    return Array(length).fill(undefined).map((_, i) => i + 1);
+/**
+ * Creates an array with numbers.
+ * @param length - The length of the desired array.
+ * @param start - The number to start counting at, can be negative, default: 1.
+ * @param step - The amount to increase in each step, can be negative, default: 1.
+ */
+export function generateNumbersArray(length: number, start = 1, step = 1): number[] {
+
+    // If step === 0
+    // Return step || 0
+
+    // If step !== 0
+    // Return ((i + (start / step)) * step) || 0
+
+    const stepper = step === 0 ? () => start || 0 : (_: undefined, index: number) => ((index + (start / step)) * step) || 0;
+    return Array(length).fill(undefined).map(stepper);
 }