+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to the array-utils project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!--## [Unreleased]-->
+## [Unreleased]
+### Added
+- Ability to give starting number and step amount to generateNumbersArray.
+
 ## [1.1.0] - 2019-05-09
 ### Added
 - Support for Node.js 12.